brain/requirements.txt
redis==5.0.1
pydantic==2.5.3
python-dotenv==1.0.0
